# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Art catalogue API
  version: 2.0.0
  description: Read-only catalogue of artworks with taxonomies, full-text search and query logging.
servers:
  - url: http://localhost:8080
paths:
...
  /api/logger:
    post:
      summary: Record a client-side event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [category, value]
              properties:
                category: {type: string, pattern: '^[A-Za-z0-9_.:-]{1,50}$'}
                value: {type: string, minLength: 1, maxLength: 1000}
      responses:
        '201': {description: Logged}
        '400': {$ref: '#/components/responses/Problem'}
        '429': {$ref: '#/components/responses/Problem'}
  /api/logs:
    get:
      summary: Read the query log (requires X-API-Key; 404 when LOGS_API_KEY is unset)
      security: [{ApiKey: []}]
      parameters:
        - {name: category, in: query, schema: {type: string}}
        - {$ref: '#/components/parameters/page'}
        - {$ref: '#/components/parameters/limit'}
      responses: {'200': {description: Log page}, '401': {$ref: '#/components/responses/Problem'}}
...
components:
  securitySchemes:
    ApiKey: {type: apiKey, in: header, name: X-API-Key}
  parameters:
    page: {name: page, in: query, schema: {type: integer, minimum: 1, maximum: 10000, default: 1}}
    limit: {name: limit, in: query, description: Values above 100 are clamped to 100, schema: {type: integer, minimum: 1, default: 10}}
...
  responses:
    Problem:
      description: RFC 9457 problem details
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
...
  schemas:
...
    Problem:
      type: object
      properties:
        type: {type: string, example: about:blank}
        title: {type: string}
        status: {type: integer}
        detail: {type: string}
        success: {type: boolean, example: false}
...